Allow MemPool late initialization
FQDN had memDataInit() in its component setup, which movedin rev 11499.
The assert is in memCheckInit() and tests that memInit() worked properly.
But if a pool is initialized only when its component is loaded, that
check will fail on several conditions unrelated to the operation of
memory. Seemingly trivial changes to component loading order is one case.

This patch allows modules to initialize/register their own pools on
demand later in the startup process.

* shuffle MEM_DONTFREE which is an existing fixed entry that must not be
  memInitCheck()'d to the end of the MemPool type enum list.

* update memCheckInit() to stop scanning for missing pools at that marker.

* shuffle pool types which are initialized by their components after the
  marker value. Such that no false problem is reported if (a) the
  component is never initialized for that worker, or (b) the component is
  only initialized during the configuration process.

* document this layout significance in the enum list to aid future pool
  additions or moves.

* add asserts to memAllocate() and memFree() to highlight the cases of
  brokenness memCheckInit() was catching. Using assert() instead of if()
  so that optimized builds can avoid the penalty of an extra test on each
  alloc/free.